feat(view): 사다리 결과를 같이 출력
