+++ src/main/java/ladder/README.md
@@ -14,7 +14,7 @@
 ### Step3
 
 - [x] 사다리 결과를 입력받는다
-- [ ] 결과를 사다리 게임 뷰에 같이 출력해야 한다
+- [x] 결과를 사다리 게임 뷰에 같이 출력해야 한다
 - [ ] 특정 참가자의 사다리 게임 결과를 계산할 수 있어야 한다
     - [ ] 결과를 보고 싶은 참가자 이름을 입력받을 수 있다
     - [ ] 결과를 출력한다
