+++ src/main/java/ladder/view/result/LadderResultView.java
@@ -1,0 +1,18 @@
+package ladder.view.result;
+
+import ladder.model.LadderResults;
+
+public class LadderResultView implements FormattableView {
+
+    private final LadderResults ladderResults;
+
+    public LadderResultView(LadderResults ladderResult) {
+        this.ladderResults = ladderResult;
+    }
+
+    @Override
+    public void render(int formatWidth) {
+        ladderResults.forEach(result -> System.out.printf("%" + formatWidth + "s", result));
+        System.out.println();
+    }
+}
