+++ src/main/java/ladder/view/result/ResultView.java
@@ -6,6 +6,6 @@ public class ResultView implements View {
 
     @Override
     public void render() {
-        System.out.println("\n실행결과\n");
+        System.out.println("\n사다리 결과\n");
     }
 }
